1428: Support issuing SSAs with software_jwks_endpoint
The `software_jwks_endpoint` is an SSA claim that contains a uri for the JWKS, this can be used to verify signatures created by the ApiClient.

Prior to this change, only the `software_jwks` claim could be set in the SSA. This claim embeds the JWKS object into the SSA. This functionality is preserved for backwards compatibility, if the caller supplies the JWKS then the `software_jwks` claim will be set and the `software_jwks_endpoint` will be omitted.

Route: `70-jwkms-apiclient-get-jwks.json` has been added and can be used to retrieve a JWKS by supplying a URI of the form: /jwkms/apiclient/jwks/$organisationId/$softwareId . The `software_jwks_endpoint` value in SSAs is a URI which calls this route.

Using SSAs with the `software_jwks_endpoint` is preferred over the old method as it how production directories (such as OB UK) are implemented. It allows the keys associated with a piece of software to be revoked or rotated without needing to issue a new SSA and complete DCR again.

Java implementations have been added by creating the module: secure-api-gateway-test-trusted-directory-ig-extensions

https://github.com/SecureApiGateway/SecureApiGateway/issues/1428