1449: Fix failing cert expiration tests
Some unit tests were failing as there is a test that the certificate
will still be valid at the start of the day in 30 days time. However,
that was failing with the date calulated being 1hr past the time at
which the cert expired. This was due to some daylight saving offset
being applied to a LocalDate that was already offset.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/1449